requirements.txt
flask
gunicorn
joblib
scikit-learn
pillow
numpy
tensorflow